+++ src/main/java/roomescape/controller/ReservationThemeController.java
@@ -12,37 +12,37 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationThemeRequest;
 import roomescape.dto.ReservationThemeResponse;
-import roomescape.service.ReservationService;
+import roomescape.service.ReservationThemeService;
 
 @RestController
 @RequestMapping("/themes")
 public class ReservationThemeController {
 
-    private final ReservationService reservationService;
+    private final ReservationThemeService reservationThemeService;
 
-    public ReservationThemeController(final ReservationService reservationService) {
-        this.reservationService = reservationService;
+    public ReservationThemeController(final ReservationThemeService reservationThemeService) {
+        this.reservationThemeService = reservationThemeService;
     }
 
     @GetMapping()
     public ResponseEntity<List<ReservationThemeResponse>> reservationThemeList() {
-        return ResponseEntity.status(HttpStatus.OK).body(reservationService.findReservationThemes());
+        return ResponseEntity.status(HttpStatus.OK).body(reservationThemeService.findReservationThemes());
     }
 
     @GetMapping("/ranking")
     public ResponseEntity<List<ReservationThemeResponse>> reservationThemeRankingList() {
-        return ResponseEntity.status(HttpStatus.OK).body(reservationService.findPopularThemes());
+        return ResponseEntity.status(HttpStatus.OK).body(reservationThemeService.findPopularThemes());
     }
 
     @PostMapping()
     public ResponseEntity<ReservationThemeResponse> reservationThemeAdd(
             @RequestBody ReservationThemeRequest request) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(reservationService.addReservationTheme(request));
+        return ResponseEntity.status(HttpStatus.CREATED).body(reservationThemeService.addReservationTheme(request));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> reservationThemeRemove(@PathVariable(name = "id") long id) {
-        reservationService.removeReservationTheme(id);
+        reservationThemeService.removeReservationTheme(id);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
